refactor: rename test for TodoItemServiceTest
